chore: Move common code into cmdhelper/
