+++ cmd/cli/evaluate/evaluate_cmd_test.go
@@ -7,8 +7,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/thomaspoignant/go-feature-flag/cmd/cli/evaluate"
-	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/helper"
 )
+
+const XAPIKeyHeader = "X-API-Key"
 
 func TestCmdEvaluate(t *testing.T) {
 	tests := []struct {
@@ -123,9 +124,9 @@ func TestCmdEvaluate(t *testing.T) {
 				"--header",
 				"Content-Type: application/json",
 				"--header",
-				helper.XAPIKeyHeader + ": 123456",
+				XAPIKeyHeader + ": 123456",
 				"--header",
-				helper.XAPIKeyHeader + ": 654321",
+				XAPIKeyHeader + ": 654321",
 				"--flag",
 				"test-flag",
 				"--ctx",
