+++ cmd/wasm/go.mod
@@ -19,7 +19,7 @@ require (
 	github.com/diegoholiveira/jsonlogic/v3 v3.9.0 // indirect
 	github.com/nikunjy/rules v1.5.0 // indirect
 	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	golang.org/x/exp v0.0.0-20240506185415-9bf2ced13842 // indirect
+	golang.org/x/exp v0.0.0-20240719175910-8a7402abbf56 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
